Move exitcode to cli
