+++ detekt-cli/src/test/kotlin/io/gitlab/arturbosch/detekt/cli/MainSpec.kt
@@ -5,6 +5,11 @@ package io.gitlab.arturbosch.detekt.cli
 import io.github.detekt.test.utils.NullPrintStream
 import io.github.detekt.test.utils.StringPrintStream
 import io.github.detekt.test.utils.resourceAsPath
+import io.github.detekt.tooling.api.InvalidConfig
+import io.github.detekt.tooling.api.IssuesFound
+import io.github.detekt.tooling.api.UnexpectedError
+import io.github.detekt.tooling.internal.DefaultAnalysisResult
+import io.github.detekt.tooling.internal.EmptyContainer
 import io.gitlab.arturbosch.detekt.cli.runners.ConfigExporter
 import io.gitlab.arturbosch.detekt.cli.runners.Runner
 import io.gitlab.arturbosch.detekt.cli.runners.VersionPrinter
@@ -72,6 +77,31 @@ class MainSpec {
             buildRunner(args, out, err)
 
             assertThat(err.toString()).isEmpty()
+        }
+    }
+
+    @Nested
+    inner class `returns different exit codes based on error` {
+
+        @Test
+        fun `returns zero on no error`() {
+            assertThat(DefaultAnalysisResult(EmptyContainer, null).exitCode()).isEqualTo(0)
+        }
+
+        @Test
+        fun `returns one on any UnexpectedError`() {
+            val unexpectedError = UnexpectedError(IllegalArgumentException())
+            assertThat(DefaultAnalysisResult(null, unexpectedError).exitCode()).isEqualTo(1)
+        }
+
+        @Test
+        fun `returns two on MaxIssuesReached`() {
+            assertThat(DefaultAnalysisResult(null, IssuesFound("")).exitCode()).isEqualTo(2)
+        }
+
+        @Test
+        fun `returns three on InvalidConfig`() {
+            assertThat(DefaultAnalysisResult(null, InvalidConfig("")).exitCode()).isEqualTo(3)
         }
     }
 }
