+++ detekt-tooling/src/main/kotlin/io/github/detekt/tooling/api/AnalysisResult.kt
@@ -8,14 +8,3 @@ interface AnalysisResult {
 
     val container: Detektion?
 }
-
-/**
- * May be used to exit the JVM via [kotlin.system.exitProcess].
- */
-@Suppress("detekt.MagicNumber")
-fun AnalysisResult.exitCode(): Int = when (error) {
-    is UnexpectedError -> 1
-    is IssuesFound -> 2
-    is InvalidConfig -> 3
-    null -> 0
-}
